Terminado funcionamiento del modelo e implementacion completa en JavaFX

+++ src/main/java/Modelo/SistemaCartas/Cartas/Carta.java
@@ -39,7 +39,7 @@ public class Carta {
         return this.puntaje.obtenerChips();
     }
 
-    public int getValorNumMultiplicador(){ return this.puntaje.obtenerMultiplicador();}
+    public float getValorNumMultiplicador(){ return this.puntaje.obtenerMultiplicador();}
 
 
     // Privados
